% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\docType{data}
\name{basic}
\alias{basic}
\title{Basic Colors}
\format{
An object of class \code{farben} (inherits from \code{character}) of length 5.
}
\usage{
basic
}
\description{
A basic color palette.
Contains the colors:
\itemize{
\item lime
\item strawberry
\item mango
\item blueberry
\item grape
}
}
\keyword{datasets}
