% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_farben.R
\name{basic_farben}
\alias{basic_farben}
\title{Basic Colors}
\usage{
basic_farben(...)
}
\arguments{
\item{...}{Specify which colors to return, by default returns all.}
}
\value{
Returns a character vector containing colors in hex code.
}
\description{
Color palette of five colors for categorical data.
}
\examples{
basic_farben()
basic_farben("strawberry", "grape")

}
