% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blind.R
\docType{data}
\name{blind}
\alias{blind}
\title{Colors for Color Blind}
\format{
An object of class \code{farben} (inherits from \code{character}) of length 5.
}
\usage{
blind
}
\description{
An accesible color palette for colorblind people. Contains 5 colors.
}
\keyword{datasets}
