% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark.R
\docType{data}
\name{dark}
\alias{dark}
\title{Dark Colors}
\format{
An object of class \code{farben} (inherits from \code{character}) of length 5.
}
\usage{
dark
}
\description{
A dark color palette.
Contains the colors:
\itemize{
\item bayern
\item bremen
\item dortmund
\item stpauli
\item schalke
}
}
\keyword{datasets}
