% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dark_farben.R
\name{dark_farben}
\alias{dark_farben}
\title{Dark Colors}
\usage{
dark_farben(...)
}
\arguments{
\item{...}{Specify which colors to return, by default returns all.}
}
\value{
Returns a character vector containing colors in hex code.
}
\description{
Color palette of five colors for categorical data.
}
\examples{
dark_farben()
dark_farben("bvb", "stpauli")

}
