% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constr.R
\name{farben}
\alias{farben}
\title{Constructor for \code{farben}}
\usage{
farben(colors)
}
\arguments{
\item{colors}{A (named) vector of hex colors.}
}
\value{
An object of class \code{farben}
}
\description{
The package \code{Farben} defines color (palettes) using S3 classes. This function creates a new instance of a class.
}
\examples{
farben(c("#683F8C", "#0f618a"))

}
