% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farben_demo.R
\name{farben_demo}
\alias{farben_demo}
\title{Plot Demo}
\usage{
farben_demo(color, include = "all", bg = "white")
}
\arguments{
\item{color}{A farben object.}

\item{include}{A character vector of which plots to include. Options are \code{"line"}, \code{"scatter"}, \code{"bar"},
or \code{"all"}.}

\item{bg}{The background color for the plot(s). Defaults to \code{"white"}.}
}
\value{
Produces 1–3 plots.
}
\description{
This function can be used to quickly assess if a color palette is useful. It returns
three different plots from the 'ChickWeight' data set.
}
\examples{
col <- farben(c("red", "blue", "yellow"))
farben_demo(color = col)

}
