% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodbad.R
\docType{data}
\name{goodbad}
\alias{goodbad}
\title{Good & Bad Colors}
\format{
An object of class \code{farben} (inherits from \code{character}) of length 7.
}
\usage{
goodbad
}
\description{
A positive / negative color palette of 6 colors. Useful for continuous data.
Use in combination with \code{colorRampPalette} for custom number of colors.
}
\examples{
# The colors themselves
goodbad

# Using a custom number of colors
colorRampPalette(goodbad)(4)

}
\keyword{datasets}
