% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitze.R
\name{hitze}
\alias{hitze}
\title{Hitze Color Palette}
\usage{
hitze(n)
}
\arguments{
\item{n}{Specifies the number of colors to return}
}
\value{
Returns a character vector of hex codes of length n.
}
\description{
This palette can be used for ordinal data.
}
\examples{
hitze(10)
}
