% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsv_adjust.R
\name{hsv_adjust}
\alias{hsv_adjust}
\title{HSV adjust}
\usage{
hsv_adjust(v.col, hue = NULL, saturation = NULL, value = NULL)
}
\arguments{
\item{v.col}{A vector of colors given in hex code.}

\item{hue}{Hue adjustment, given as -100 to 100.}

\item{saturation}{Saturation adjustment, given as -100 to 100.}

\item{value}{Value adjustment, given as -100 to 100.}
}
\value{
Returns an adjusted vector of colors given in hex code.
}
\description{
This function is used to quickly adjust the HSV values of colors.
HSV stands for hue, saturation, and value.
}
\details{
The adjustments are linear interpolations of the current hsv value and the maximum/minimum
values.
}
\examples{
hsv_adjust(c("red", "blue", "green"), hue = 10)
hsv_adjust(c("red", "blue", "green"), saturation = -20, value = -10)
}
