% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsv_show.R
\name{hsv_show}
\alias{hsv_show}
\title{HSV Show}
\usage{
hsv_show(v.col, dec = 3)
}
\arguments{
\item{v.col}{A vector of colors given in hex code.}

\item{dec}{Number of decimals to be shown. Defaults to three.}
}
\value{
Returns a matrix with three rows and a column for each given color.
}
\description{
This function shows the HSV values for given colors.
HSV stands for hue, saturation, and value. This function is best used in combination
with the hsv_adjust function to tune a given color palette.
}
\examples{
hsv_show(c("red", "blue", "yellow"))

}
