% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.farben}
\alias{plot.farben}
\title{Plot Colors}
\usage{
\method{plot}{farben}(x, ...)
}
\arguments{
\item{x}{Colors that should be plotted. Given as a \code{farben} object.}

\item{...}{Additional arguments passed to plot.}
}
\value{
Plot containing all given colors, including their hex code.
}
\description{
This \code{farben} method for \code{plot} offers a quick visualization of colors.
}
