% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_colors.R
\name{plot_colors}
\alias{plot_colors}
\title{Plot Colors}
\usage{
plot_colors(v.col)
}
\arguments{
\item{v.col}{Colors that should be plotted. Given as a vector of hex codes.}
}
\value{
Plot containing all given colors, including their hex code.
}
\description{
The function offers a quick visualization of colors. Can be used in combination
with color adjustment functions to find a set of colors that suits you best.
}
\examples{
plot_colors(c("red", "blue", "yellow"))

}
