% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_demo.R
\name{plot_demo}
\alias{plot_demo}
\title{Plot Demo}
\usage{
plot_demo(color, include = "all")
}
\arguments{
\item{color}{A vector of colors given in hex code.}

\item{include}{A character vector of which plots to include.}
}
\value{
Return 1-3 plots.
}
\description{
This function can be used to quickly asses if a color palette is useful. It returns
three different plots from the 'ChickWeight' data set.
}
\examples{
plot_demo(color = c("red", "blue", "yellow"))
}
