% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbow.R
\docType{data}
\name{rainbow}
\alias{rainbow}
\title{Rainbow Colors}
\format{
An object of class \code{farben} (inherits from \code{character}) of length 8.
}
\usage{
rainbow
}
\description{
A large color palette. Contains 8 colors.
}
\keyword{datasets}
