% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/royal.R
\docType{data}
\name{royal}
\alias{royal}
\title{Royal Colors}
\format{
An object of class \code{farben} (inherits from \code{character}) of length 3.
}
\usage{
royal
}
\description{
An "old-money" color palette. Contains 3 colors.
}
\keyword{datasets}
