% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\docType{data}
\name{space}
\alias{space}
\title{Space Colors}
\format{
An object of class \code{farben} (inherits from \code{character}) of length 3.
}
\usage{
space
}
\description{
A spacy color palette. Contains 3 colors.
}
\keyword{datasets}
