% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warmth_adjust.R
\name{warmth_adjust}
\alias{warmth_adjust}
\title{Warmth Adjust}
\usage{
warmth_adjust(v.col, adjust)
}
\arguments{
\item{v.col}{A vector of colors given in hex code.}

\item{adjust}{The adjustment, given from -100 to 100. Positive = warmer.}
}
\value{
Returns the adjusted colors
}
\description{
This function adjusts the "warmth" of a given set of colors. This is done by shifting the
colors to red for warm and to blue for cold.
}
\examples{
warmth_adjust(c("green", "purple"), adjust = 30)
}
