% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodbad.R
\name{goodbad}
\alias{goodbad}
\title{Good Bad Color Palette}
\usage{
goodbad(n)
}
\arguments{
\item{n}{Specifies the number of colors to return}
}
\value{
Returns a character vector of hex codes of length n.
}
\description{
This color palette can be used for ordinal data. It is of best use when the data has a notion of
better or worse.
}
\examples{
goodbad(10)
}
