% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitze.R
\docType{data}
\name{hitze}
\alias{hitze}
\title{Hitze Colors}
\format{
An object of class \code{farben} (inherits from \code{character}) of length 6.
}
\usage{
hitze
}
\description{
A heat color palette of 6 colors. Useful for continuous data.
Use in combination with \code{colorRampPalette} for custom number of colors.
}
\examples{
# The colors themselves
hitze

# Using a custom number of colors
colorRampPalette(hitze)(5)

}
\keyword{datasets}
